+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- change clients db to postgres. [#187](https://github.com/DevAlexandreCR/gorda-admin/pull/187)
+
 # Release Notes for 1.5.x
 
 ## [1.5.15(2025-11-12)](https://github.com/DevAlexandreCR/admin-driver/compare/1.5.15...1.5.14)